Huge refactor, fixed bug in run_keys
A lot has changed. Examples are now wrapped in an Example Collection, which helps handling the labeling of examples. A special object for handling the BackgoundKnowledge has been made, since there can be labeling rules in the background knowledge. More effort has been put into making sure examples are labeled when they have to be. Examples are now wrapped in an ExampleWrapper. This makes the fields an ExampleWrapper object can have explicit, instead of dynamically adding  values to SimpleProgram and ClauseDB objects. That was how it used to be done, which could give AttributeErrors which had to be catched. ExampleWrapper is also important because it contains a value classification_term. This term is the prolog statement that might be used to label an example. Special treatment of this statement is necessary, because if we keep it in the regular example, classifying the example with the constructed decision tree would have no use, since the prolog example contains a labeling term. This labeling term would always give the correct label. Using the Muta example with run_keys now shows how the decision tree now correcly misclassifies the examples which it could not cover in the learning process.