Bugfix in SimpleProgramExampleWrapper
Bug made it so that a SimpleProgramExampleWrapper on construction always contained an empty logic program, even when given another logic program. 
IMPORTANT: functions are first-class citizens in Python. Never use a mutable default object as assignment in the construtor argument list, instead if you want all function calls to share this object. Closures!